Move more l10n logic from app/factory.py to app/l10n.py.
