+++ app/tests/test_l10n.py
@@ -1,7 +1,15 @@
 import os
+from flask import Flask
+from nose.tools import eq_
 
 from .test_views import ViewTestCase
 from app import l10n
+
+def test_flask_security_strings_do_not_interpolate():
+    app = Flask('test')
+    l10n.configure_app(app)
+    eq_(unicode(app.config['SECURITY_MSG_EMAIL_ALREADY_ASSOCIATED'][0]),
+        '%(email)s is already associated with an account.')
 
 class LocalizationTests(ViewTestCase):
     def test_all_translations_exist(self):
